refactor(attach): slim trap uprobe backend and add ISA comments
Consolidate three small test files into test_trap_misc.cpp, remove
machine-specific benchmark results.md, replace inline /proc/self/maps
parsing with get_module_base_addr(), inline the no-op prepare_thread(),
compress branch switch cases to single lines, and add English comments
to RISC-V instruction decode/emulate, RCU publish pattern, and signal
dispatch logic.  Also fix JIT compilation failure to fall back to the
interpreter instead of returning an error.

Net reduction: ~275 lines from the trap uprobe diff.
